SqlForge-Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
aiosqlite
typing-extensions
python-multipart
pytest
httpx
google-genai
psycopg2-binary
pymysql
pymssql
oracledb
redis
pymongo
sqlglot
pysqlsync
sshtunnel
apscheduler
pandas
openpyxl
dbfread
dbf
pyodbc
apscheduler
pyinstaller